requirements.txt
python-dotenv
pre-commit
supertriplets
fire
matplotlib
pyarrow
faiss-gpu
streamlit